requirements.txt
diffusers
transformers
torch
torchvision
deepface
IPython
Pillow
matplotlib
numpy
